requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

pandas == 1.2.5
python_slugify == 8.0.1
selenium == 4.6.0
streamlit == 1.11.0
streamlit_option_menu == 0.3.2
